pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools"]

[project]
name = "gh-comment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "PyGithub==2.5.0"
]
[project.scripts]
gh-comment = "gh_comment.main:main"
